A forgotten cast in the rar and par settings made it so that any size config over 2 GB (Int32.MaxValue) wrapped around.

+++ Util/Configuration/RarNParSetting.cs
@@ -16,7 +16,7 @@ namespace Util.Configuration
 
         public Int64 FromSizeBytes
         {
-            get { return FromSize*1024*1024; }
+            get { return (Int64)FromSize*1024*1024; }
         }
 
         [DataMember(Order = 1)]
